requirements.txt
Flask==3.1.0
gunicorn==
langchain-ollama==0.3.4
python-dotenv==1.0.1
langgraph==0.3.21
gunicorn==23.0.0